Add flux period estimation function.
Most of my floppy drive images consist of multiple runs around the
disk, so it seems reasonable that we should be able to line these up
to determine where weak flux transitions exist, to correct errors.

I'm thinking about just how to do it; the flux period estimator
might not be the right way to do it, but I'll have to experiment with
different approaches.

(sector_data also needs to be cleaned up properly, and there are some
other TODOS to well, do.)