Add flux filename as input parameter, change ordinal search clock inference.
The user can now specify what file to analyze, and the clock inference
should work better. In addition, it will return -2 for fluxes where the
clock bands were guessed correctly but it's impossible to fit a single
clock value to the result. This may be used for explicit band specification
later.