Add very quick and dirty fix for 111_clock_failure_1, and find new bug.
111_clock_failure_1 is due to an overlap in search match positions.
Currently the program just checks if the distance to the next match
is so short that it can't be a full match, but this is susceptible
to more complex bugs (e.g. A1A1A1A1F8, then the first A1A1A1 match
would be too short but would also stop the program from finding the
true match). Fix later.

And the new error (exercised by 111_clock_failure_2) is really hard
to debug.