Fix the "overlapping substring" problem by adding a high bit to preambles.
Remove offset_sequence code from preambles as it's no longer needed;
improved ordinal search logic makes it redundant.

The overlapping substring problem is where a preamble is too long, e.g.
A1A1A1A1FB, and then the detector finds the first three A1s; the
chunk logic then considers the first chunk to be a preamble, cutting
off the A1s from the next chunk. Then the *true* preamble (A1A1A1FB)
is not recognized because the next chunk contains only A1FB.

This is mostly fixed by making the search pattern non-overlapping by
nature; in this case, by appending a 01 to the MFM train search pattern.
This is correct because every preamble (A1 or C2) has a high bit (actually
a high nibble) on the byte following it. Yet A1A1A1F* is not a substring
of itself, hence the overlapping substring problem vanishes.

It might still occur if there's a false positive that claims some bytes
from a true positive, but I think that's very unlikely. I'll deal with
it when I find an appropriate test.