decoder: Enable linking IDAM (metadata) to DAM (data) after padding separation.
flux-analyze used to not be able to link IDAM to DAM if an earlier run had
split the IDAM timeslice (usually separating out the padding following the
IDAM itself). Now it just checks if it's possible for a DAM to fit between
the last IDAM and the current DAM; if not, no other data could've been there,
so matching the DAM to the IDAM is all right.

This could cause a problem with nonstandard floppy layouts, e.g. suppose
first IDAM X appears then a corrupted IDAM (Y) that's not recognized, then
the DAM itself. It will be assigned to X but should be unassigned. Conversely
suppose first we have IDAM X then a false positive IDAM Y with bad CRC;
it'll attach the DAM to Y instead of to X.