Add brute-force dewarping.
main.cc now does a run of different values of alpha (smoothing parameter)
when decoding. This is unusually effective for such a simple change,
but also quite slow. It's intended mostly as a proof of concept.

It has also uncovered some bugs; because it's been so long since I've
last dealt with this code, I'm not sure where the bug resides, though
I'm inclined to blame the fiddly nature of the timeline construct. It
might need a more thorough redesign.