Start working on some dynamic programming experiments.
The DP approach should allow the clock to change throughout the stream,
while following MFM format constraints. If it works, it should make
decoding *much* easier and fix the bootstrap problem where you
need to know the initial clock to figure out what data is encoded.

But more has to be done before I can check if it works.